Indexes and Rollup_Join Docs Added
I've added documentation for indexes and rollup_join and updated everything according to Tom’s feedback.

There’s one intentional deviation: in the index example, the pre-aggregation includes the dimensions id and name. You had suggested keeping only size, but I decided to retain name because it seemed more meaningful for the example, and I kept id since it’s used in the join with orders in the rollup_join section below. That’s also why the index is created on id. So we can saw its pre-aggregated at id,name level.